# services/trading-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing portfolios, executing trades, and handling financial market operations.",
        "title": "Trading Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/listings/forex": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "List forex pairs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by base/quote",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min rate",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max rate",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc|desc",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedForexResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/listings/futures": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "List futures contracts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by ticker or name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exchange MIC prefix",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min price",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max price",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min ask",
                        "name": "ask_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max ask",
                        "name": "ask_max",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min bid",
                        "name": "bid_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max bid",
                        "name": "bid_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Min volume",
                        "name": "volume_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max volume",
                        "name": "volume_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by settlement date (YYYY-MM-DD)",
                        "name": "settlement_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "price|volume|maintenance_margin",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc|desc",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedFuturesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/listings/options": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "List options",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by ticker or name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exchange MIC prefix",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min price",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max price",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min ask",
                        "name": "ask_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max ask",
                        "name": "ask_max",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min bid",
                        "name": "bid_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max bid",
                        "name": "bid_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Min volume",
                        "name": "volume_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max volume",
                        "name": "volume_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by settlement date (YYYY-MM-DD)",
                        "name": "settlement_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "price|volume|maintenance_margin",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc|desc",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedOptionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/listings/stock/listingId": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves stock information for a specific listing by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Get stock details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Listing ID",
                        "name": "listingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.StockDetailedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/listings/stocks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "List stocks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by ticker or name",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exchange MIC prefix",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min price",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max price",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min ask",
                        "name": "ask_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max ask",
                        "name": "ask_max",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Min bid",
                        "name": "bid_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Max bid",
                        "name": "bid_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Min volume",
                        "name": "volume_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max volume",
                        "name": "volume_max",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "price|volume|maintenance_margin",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc|desc",
                        "name": "sort_dir",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginatedStockResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.DailyPriceResponse": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "number"
                },
                "bid": {
                    "type": "number"
                },
                "change": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ForexResponse": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "number"
                },
                "base": {
                    "type": "string"
                },
                "bid": {
                    "type": "number"
                },
                "change": {
                    "type": "number"
                },
                "forexPairId": {
                    "type": "integer"
                },
                "initialMarginCost": {
                    "type": "number"
                },
                "maintenanceMargin": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "quote": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
        "dto.FuturesResponse": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "number"
                },
                "bid": {
                    "type": "number"
                },
                "change": {
                    "type": "number"
                },
                "contract_size": {
                    "type": "number"
                },
                "contract_unit": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                },
                "initial_margin_cost": {
                    "type": "number"
                },
                "listing_id": {
                    "type": "integer"
                },
                "maintenance_margin": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "settlement_date": {
                    "type": "string"
                },
                "ticker": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
        "dto.OptionResponse": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "number"
                },
                "bid": {
                    "type": "number"
                },
                "change": {
                    "type": "number"
                },
                "exchange": {
                    "type": "string"
                },
                "implied_volatility": {
                    "type": "number"
                },
                "initial_margin_cost": {
                    "type": "number"
                },
                "listing_id": {
                    "type": "integer"
                },
                "maintenance_margin": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "open_interest": {
                    "type": "integer"
                },
                "option_type": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "settlement_date": {
                    "type": "string"
                },
                "strike": {
                    "type": "number"
                },
                "ticker": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PaginatedForexResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ForexResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.PaginatedFuturesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FuturesResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.PaginatedOptionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OptionResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "dto.PaginatedStockResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StockResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.StockDetailedResponse": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "number"
                },
                "bid": {
                    "type": "number"
                },
                "change": {
                    "type": "number"
                },
                "dividend_yield": {
                    "type": "number"
                },
                "exchange": {
                    "type": "string"
                },
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DailyPriceResponse"
                    }
                },
                "initial_margin_cost": {
                    "type": "number"
                },
                "listing_id": {
                    "type": "integer"
                },
                "maintenance_margin": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.OptionResponse"
                    }
                },
                "outstanding_shares": {
                    "type": "number"
                },
                "price": {
                    "type": "number"
                },
                "ticker": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
        "errors.AppError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}